Deconstruct channels class
